refactor: 패키지 구조 수정 및 검증 예외처리 일관적 처리
