+++ src/main/java/com/goggles/mentoring_service/MentoringServerApplication.java
@@ -2,7 +2,9 @@ package com.goggles.mentoring_service;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
+@EnableJpaAuditing
 @SpringBootApplication
 public class MentoringServerApplication {
