--- src/main/java/com/goggles/mentoring_service/domain/BookingType.java
@@ -1,5 +1,0 @@
-package com.goggles.mentoring_service.domain;
-
-public enum BookingType {
-	ONE_TIME, SELF_SELECT, AUTO_REPEAT
-}
