--- src/main/java/com/goggles/mentoring_service/domain/MentoringRepository.java
@@ -1,9 +1,0 @@
-package com.goggles.mentoring_service.domain;
-
-import java.util.Optional;
-import java.util.UUID;
-
-public interface MentoringRepository {
-	void save(Mentoring mentoring);
-	Optional<Mentoring> findById(MentoringId id);
-}
