--- src/main/java/com/goggles/mentoring_service/domain/MentoringStatus.java
@@ -1,6 +1,0 @@
-package com.goggles.mentoring_service.domain;
-
-public enum MentoringStatus {
-	ACTIVE,
-	INACTIVE
-}
