--- src/main/java/com/goggles/mentoring_service/domain/MentoringType.java
@@ -1,6 +1,0 @@
-package com.goggles.mentoring_service.domain;
-
-public enum MentoringType {
-	ONE_ON_ONE,
-	GROUP
-}
