--- src/main/java/com/goggles/mentoring_service/domain/SessionStatus.java
@@ -1,7 +1,0 @@
-package com.goggles.mentoring_service.domain;
-
-public enum SessionStatus {
-	AVAILABLE,
-	BOOKED,
-	NOT_AVAILABLE
-}
