--- src/main/java/com/goggles/mentoring_service/domain/UserType.java
@@ -1,5 +1,0 @@
-package com.goggles.mentoring_service.domain;
-
-public enum UserType {
-	STUDENT, INSTRUCTOR, MASTER
-}
