+++ src/main/java/com/goggles/mentoring_service/domain/common/MentoringDomainException.java
@@ -1,0 +1,8 @@
+package com.goggles.mentoring_service.domain.common;
+
+public abstract class MentoringDomainException extends RuntimeException {
+
+	protected MentoringDomainException(String message) {
+		super(message);
+	}
+}
