+++ src/main/java/com/goggles/mentoring_service/domain/common/MentoringValidationException.java
@@ -1,0 +1,8 @@
+package com.goggles.mentoring_service.domain.common;
+
+public class MentoringValidationException extends MentoringDomainException {
+
+	public MentoringValidationException(String message) {
+		super(message);
+	}
+}
