+++ src/main/java/com/goggles/mentoring_service/domain/common/UserType.java
@@ -1,0 +1,5 @@
+package com.goggles.mentoring_service.domain.common;
+
+public enum UserType {
+	STUDENT, INSTRUCTOR, MASTER
+}
