--- src/main/java/com/goggles/mentoring_service/domain/exception/BookedSessionStatusCannotBeChangedException.java
@@ -1,10 +1,0 @@
-package com.goggles.mentoring_service.domain.exception;
-
-import com.goggles.common.exception.ConflictException;
-
-public class BookedSessionStatusCannotBeChangedException extends ConflictException {
-
-	public BookedSessionStatusCannotBeChangedException() {
-		super("예약된 세션은 상태를 변경할 수 없습니다.");
-	}
-}
