--- src/main/java/com/goggles/mentoring_service/domain/exception/InvalidMentoringPolicyException.java
@@ -1,10 +1,0 @@
-package com.goggles.mentoring_service.domain.exception;
-
-
-public class InvalidMentoringPolicyException extends MentoringValidationException {
-
-	public InvalidMentoringPolicyException(String message) {
-		super(message);
-	}
-}
-
