--- src/main/java/com/goggles/mentoring_service/domain/exception/MentoringDomainException.java
@@ -1,8 +1,0 @@
-package com.goggles.mentoring_service.domain.exception;
-
-public abstract class MentoringDomainException extends RuntimeException {
-
-	protected MentoringDomainException(String message) {
-		super(message);
-	}
-}
