--- src/main/java/com/goggles/mentoring_service/domain/exception/MentoringValidationException.java
@@ -1,8 +1,0 @@
-package com.goggles.mentoring_service.domain.exception;
-
-public class MentoringValidationException extends MentoringDomainException {
-
-	public MentoringValidationException(String message) {
-		super(message);
-	}
-}
