+++ src/main/java/com/goggles/mentoring_service/domain/mentoring/BookingType.java
@@ -1,0 +1,5 @@
+package com.goggles.mentoring_service.domain.mentoring;
+
+public enum BookingType {
+	ONE_TIME, SELF_SELECT, AUTO_REPEAT
+}
