+++ src/main/java/com/goggles/mentoring_service/domain/mentoring/MentoringStatus.java
@@ -1,0 +1,6 @@
+package com.goggles.mentoring_service.domain.mentoring;
+
+public enum MentoringStatus {
+	ACTIVE,
+	INACTIVE
+}
