+++ src/main/java/com/goggles/mentoring_service/domain/mentoring/MentoringType.java
@@ -1,0 +1,6 @@
+package com.goggles.mentoring_service.domain.mentoring;
+
+public enum MentoringType {
+	ONE_ON_ONE,
+	GROUP
+}
