+++ src/main/java/com/goggles/mentoring_service/domain/mentoring/SessionStatus.java
@@ -1,0 +1,7 @@
+package com.goggles.mentoring_service.domain.mentoring;
+
+public enum SessionStatus {
+	AVAILABLE,
+	BOOKED,
+	NOT_AVAILABLE
+}
