+++ src/main/java/com/goggles/mentoring_service/domain/mentoring/exception/InvalidTimeRangeException.java
@@ -1,0 +1,10 @@
+package com.goggles.mentoring_service.domain.mentoring.exception;
+
+import com.goggles.mentoring_service.domain.common.MentoringValidationException;
+
+public class InvalidTimeRangeException extends MentoringValidationException {
+
+	public InvalidTimeRangeException(String message) {
+		super(message);
+	}
+}
