+++ src/main/java/com/goggles/mentoring_service/domain/mentoring/repository/MentoringRepository.java
@@ -1,0 +1,11 @@
+package com.goggles.mentoring_service.domain.mentoring.repository;
+
+import com.goggles.mentoring_service.domain.mentoring.Mentoring;
+import com.goggles.mentoring_service.domain.mentoring.MentoringId;
+
+import java.util.Optional;
+
+public interface MentoringRepository {
+	void save(Mentoring mentoring);
+	Optional<Mentoring> findById(MentoringId id);
+}
